singleAgent/requirements.txt
openai
python-dotenv
tiktoken
rich
prompt_toolkit